Updates preservation.gemspec
We were unintentionally overwriting rake tasks in the host app with rake tasks
from the gem. This highlighted the bigger problem of including too many files
as part of the installed gem. So we added a bit of logic to the .gemspec to
exclude files that are only needed when developing the gem.